XC-412: Move more type conversions to separate files
